Lesson5. Active push
Удалено все лишнее
Добавил запросы из Postman
