--- WorkManager/WorkManager/DAL/Interfaces/IClient.cs
@@ -1,17 +1,0 @@
-﻿using WorkManager.DAL.Models;
-
-namespace WorkManager.DAL.Interfaces.Archive
-{
-    public interface IClient
-    {
-        /// <summary>
-        /// Контракт клиента
-        /// </summary>
-        //public ClientContract  ClientContract { get; }
-
-        /// <summary>
-        /// Фирма клиента
-        /// </summary>
-        public string Company { get; }
-    }
-}
