--- WorkManager/WorkManager/DAL/Interfaces/IEmployee.cs
@@ -1,17 +1,0 @@
-﻿using System;
-
-namespace WorkManager.DAL.Interfaces.Archive
-{
-    public interface IEmployee
-    {
-        /// <summary>
-        /// Размер з.п. сотрудника за 1 час
-        /// </summary>
-        public int HourSalary { get; }
-
-        /// <summary>
-        /// Потраченное время сотрудника
-        /// </summary>
-        public TimeSpan SpendingTime { get; }
-    }
-}
