--- WorkManager/WorkManager/DAL/Models/ClientContract.cs
@@ -1,10 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations.Schema;
-
-namespace WorkManager.DAL.Models.Archive
-{
-    [Table("ClientContracts", Schema = "WorkManager")]
-    public sealed class ClientContract : WorkResultEntity
-    {
-
-    }
-}
