+++ WorkManager/WorkManager/Repositories/ClientsRepository.cs
@@ -1,55 +1,54 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
-using WorkManager.DAL.Repositories.Interfaces;
+using WorkManager.Repositories.Interfaces;
 using WorkManager.Data.Contexts;
 using WorkManager.Data.Models;
 
-namespace WorkManager.DAL.Repositories
+namespace WorkManager.Repositories
 {
-    internal sealed class EmployeesRepository : IRepository<int, Employee>
+    internal sealed class ClientsRepository : IRepository<int, Client>
     {
         /// <summary>
         /// Контекст БД
         /// </summary>
         private readonly WorkManagerDbContext _context;
 
-
-        public EmployeesRepository(WorkManagerDbContext dbContext)
+        public ClientsRepository(WorkManagerDbContext dbContext)
         {
             _context = dbContext;
         }
 
-        public bool Create(Employee entity)
+        public bool Create(Client entity)
         {
-            _context.Employees.Add(entity);
+            _context.Clients.Add(entity);
             _context.SaveChanges();
             return true;
         }
 
-        public IReadOnlyDictionary<int, Employee> Get()
+        public IReadOnlyDictionary<int, Client> Get()
         {
             // Подгружаю элементы из БД
-            IEnumerable<Employee> entityColl = _context
-                .Employees
+            IEnumerable<Client> entityColl = _context
+                .Clients
                 .Where(c => c.IsDeleted == false)
                 .AsEnumerable();
 
-            Dictionary<int, Employee> entitysIndex = entityColl
+            Dictionary<int, Client> entitysIndex = entityColl
                 .ToDictionary(c => c.Id, c => c);
 
-            IReadOnlyDictionary<int, Employee> result = entitysIndex;
+            IReadOnlyDictionary<int, Client> result = entitysIndex;
             return result;
         }
 
-        public Employee GetById(int id)
+        public Client GetById(int id)
         {
-            return _context.Employees.SingleOrDefault(c => c.Id == id);
+            return _context.Clients.SingleOrDefault(c => c.Id == id);
         }
 
         public bool UpdateById(int id, string reqColumnName, string value)
         {
-            Employee entity = GetById(id);
+            Client entity = GetById(id);
             if (entity != null)
             {
                 PropertyInfo prop = entity.GetType().GetProperty(reqColumnName, BindingFlags.Public | BindingFlags.Instance);
@@ -68,7 +67,7 @@ namespace WorkManager.DAL.Repositories
 
         public bool DeleteById(int id)
         {
-            Employee entity = GetById(id);
+            Client entity = GetById(id);
             if (entity != null)
             {
                 entity.IsDeleted = true;
