+++ WorkManager/WorkManager/Repositories/EmployeesRepository.cs
@@ -1,54 +1,55 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
-using WorkManager.DAL.Repositories.Interfaces;
+using WorkManager.Repositories.Interfaces;
 using WorkManager.Data.Contexts;
 using WorkManager.Data.Models;
 
-namespace WorkManager.DAL.Repositories
+namespace WorkManager.Repositories
 {
-    internal sealed class InvoicesRepository : IRepository<int, Invoice>
+    internal sealed class EmployeesRepository : IRepository<int, Employee>
     {
         /// <summary>
         /// Контекст БД
         /// </summary>
         private readonly WorkManagerDbContext _context;
 
-        public InvoicesRepository(WorkManagerDbContext dbContext)
+
+        public EmployeesRepository(WorkManagerDbContext dbContext)
         {
             _context = dbContext;
         }
 
-        public bool Create(Invoice entity)
+        public bool Create(Employee entity)
         {
-            _context.Invoices.Add(entity);
+            _context.Employees.Add(entity);
             _context.SaveChanges();
             return true;
         }
 
-        public IReadOnlyDictionary<int, Invoice> Get()
+        public IReadOnlyDictionary<int, Employee> Get()
         {
             // Подгружаю элементы из БД
-            IEnumerable<Invoice> entityColl = _context
-                .Invoices
+            IEnumerable<Employee> entityColl = _context
+                .Employees
                 .Where(c => c.IsDeleted == false)
                 .AsEnumerable();
 
-            Dictionary<int, Invoice> entitysIndex = entityColl
+            Dictionary<int, Employee> entitysIndex = entityColl
                 .ToDictionary(c => c.Id, c => c);
 
-            IReadOnlyDictionary<int, Invoice> result = entitysIndex;
+            IReadOnlyDictionary<int, Employee> result = entitysIndex;
             return result;
         }
 
-        public Invoice GetById(int id)
+        public Employee GetById(int id)
         {
-            return _context.Invoices.SingleOrDefault(c => c.Id == id);
+            return _context.Employees.SingleOrDefault(c => c.Id == id);
         }
 
         public bool UpdateById(int id, string reqColumnName, string value)
         {
-            Invoice entity = GetById(id);
+            Employee entity = GetById(id);
             if (entity != null)
             {
                 PropertyInfo prop = entity.GetType().GetProperty(reqColumnName, BindingFlags.Public | BindingFlags.Instance);
@@ -67,7 +68,7 @@ namespace WorkManager.DAL.Repositories
 
         public bool DeleteById(int id)
         {
-            Invoice entity = GetById(id);
+            Employee entity = GetById(id);
             if (entity != null)
             {
                 entity.IsDeleted = true;
