+++ WorkManager/WorkManager/Repositories/UsersRepository.cs
@@ -1,10 +1,10 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
-using WorkManager.DAL.Repositories.Interfaces;
+using WorkManager.Repositories.Interfaces;
 using WorkManager.Data.Contexts;
 using WorkManager.Data.Models;
 
-namespace WorkManager.DAL.Repositories
+namespace WorkManager.Repositories
 {
     internal sealed class UsersRepository : IUserRepository
     {
