+++ WorkManager/WorkManager/Responses/InvoiceResponse.cs
@@ -2,7 +2,7 @@
 using System;
 using System.Collections.Generic;
 using WorkManager.Data.Models;
-using WorkManager.DAL.Repositories.Interfaces;
+using WorkManager.Repositories.Interfaces;
 using WorkManager.Responses.Interfaces;
 
 namespace WorkManager.Responses
