+++ WorkManager/WorkManager/Startup.cs
@@ -10,8 +10,8 @@ using Microsoft.IdentityModel.Tokens;
 using Microsoft.OpenApi.Models;
 using System;
 using System.Text;
-using WorkManager.DAL.Repositories;
-using WorkManager.DAL.Repositories.Interfaces;
+using WorkManager.Repositories;
+using WorkManager.Repositories.Interfaces;
 using WorkManager.Data.Contexts;
 using WorkManager.Data.Models;
 using WorkManager.MySQLsettings;
